+++ orchestration/src/main/java/com/sap/ai/sdk/orchestration/OrchestrationConfigReference.java
@@ -4,13 +4,14 @@ import com.google.common.annotations.Beta;
 import javax.annotation.Nonnull;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import lombok.Value;
 
-// JONAS: Making this a sealed interface permitting 2 classes for ID and SNV seemed like an overkill
-// JONAS: Not a record because I do not want a public all-args constructor
-// JONAS: Is string correct for ID? (Use UUID?)
-// JONAS: Split SNV into 3 methods?
-
+/**
+ * Class representing a reference to an Orchestration config stored in prompt registry.
+ *
+ * @since 1.14.0
+ */
 @Value
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Beta
@@ -20,12 +21,72 @@ public class OrchestrationConfigReference {
   String name;
   String version;
 
+  /**
+   * Build a reference from an ID.
+   *
+   * @param id The id of the reference
+   * @return A reference object with the specified id
+   * @since 1.14.0
+   */
+  @Nonnull
   public static OrchestrationConfigReference fromId(@Nonnull final String id) {
     return new OrchestrationConfigReference(id, null, null, null);
   }
 
-  public static OrchestrationConfigReference fromScenarioNameVersion(
-      @Nonnull final String scenario, @Nonnull final String name, @Nonnull final String version) {
-    return new OrchestrationConfigReference(null, scenario, name, version);
+  /**
+   * Build a reference from a scenario, name, and version.
+   *
+   * @param scenario The scenario of the reference
+   * @return A builder object with the specified scenario
+   * @since 1.14.0
+   */
+  @Nonnull
+  public static Builder fromScenario(@Nonnull final String scenario) {
+    return new Builder(scenario);
+  }
+
+  /**
+   * Builder to create an Orchestration config reference from scenario, name, and version.
+   *
+   * @since 1.14.0
+   */
+  @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
+  public static class Builder {
+    private final String scenario;
+
+    /**
+     * Build a reference from a scenario, name, and version.
+     *
+     * @param name The name of the reference
+     * @return A builder object with the specified scenario and name
+     * @since 1.14.0
+     */
+    @Nonnull
+    public Builder1 name(@Nonnull final String name) {
+      return new Builder1(scenario, name);
+    }
+  }
+
+  /**
+   * Builder to create an Orchestration config reference from scenario, name, and version.
+   *
+   * @since 1.14.0
+   */
+  @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
+  public static class Builder1 {
+    private final String scenario;
+    private final String name;
+
+    /**
+     * Build a reference from a scenario, name, and version.
+     *
+     * @param version The version of the reference
+     * @return A reference object with the specified scenario, name, and version
+     * @since 1.14.0
+     */
+    @Nonnull
+    public OrchestrationConfigReference version(@Nonnull final String version) {
+      return new OrchestrationConfigReference(null, scenario, name, version);
+    }
   }
 }